requirements.txt
ipython==8.12.3
matplotlib==3.10.6
numpy==2.3.3
rdkit==2024.9.3
scikit_learn==1.7.2
torch==1.12.0+cu113
torch_geometric==2.6.1
torch_scatter==2.0.9
tqdm==4.66.5
umap_learn==0.5.7
